requirements.txt
Django==3.0.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
pytz
gunicorn
django-cors-headers
Pillow
psycopg2-binary